refactor{module}:rename module to follow PEP8 standard and added __init__.py file to the module
